chore: update reviewer comments
