+++ src/CarbonSDK.ts
@@ -576,7 +576,6 @@ class CarbonSDK {
 
     const wallet = CarbonWallet.withRainbowKit(rainbowKit, evmChainId, rainbowKitWalletOpts.publicKeyBase64, addressOptions, rainbowKitWalletOpts.walletProvider, {
       ...opts,
-      isRainbowKit: true,
       network: this.network,
       config: this.configOverride,
     })
