--- src/types/index.d.ts
@@ -1,6 +1,0 @@
-
-declare module "ledger-cosmos-js";
-
-interface WindowEventMap {
-  "eip6963:announceProvider": CustomEvent
-}
