+++ src/util/types.d.ts
@@ -1,2 +1,7 @@
 declare module "base58check";
 declare module "secp256r1";
+declare module "ledger-cosmos-js";
+
+interface WindowEventMap {
+  "eip6963:announceProvider": CustomEvent
+}
